app/requirements.txt
python-telegram-bot
requests
python-dotenv==1.0.1
pytz
apscheduler==3.9.1.post1
